feat: Move useGaletoOraclePrice to apps/web
