+++ apps/web/package.json
@@ -122,7 +122,8 @@
     "uuid": "^8.0.0",
     "viem": "1.19.11",
     "wagmi": "1.4.7",
-    "zod": "^3.22.3"
+    "zod": "^3.22.3",
+    "@pythnetwork/pyth-evm-js": "^1.29.2"
   },
   "devDependencies": {
     "@angleprotocol/sdk": "^3.0.121",
