+++ apps/web/src/views/Predictions/hooks/useGaletoOraclePrice.ts
