+++ packages/prediction/package.json
@@ -14,8 +14,6 @@
     "@pancakeswap/chains": "workspace:*",
     "@pancakeswap/tokens": "workspace:*",
     "@pancakeswap/pools": "workspace:*",
-    "@pythnetwork/pyth-evm-js": "^1.29.2",
-    "@tanstack/react-query": "^4.28.0",
     "viem": "1.19.11",
     "wagmi": "1.4.7"
   },
